♻refactor: using latest ansible module to be more future-proof and support latest ansible versions
Ansible-lint:
```
Include has some unintuitive behaviours depending on if it is running in a static or dynamic in play or in playbook context, in an effort to clarify behaviours we are moving to a new set modules (ansible.builtin.include_tasks, ansible.builtin.include_role, ansible.builtin.import_playbook, ansible.builtin.import_tasks) that have well established and clear behaviours.
This module will still be supported for some time but we are looking at deprecating it in the near future.
```

Fix: I have decided to go for ansible include_tasks module because it is more versatile and on a par with this module (for more info about differences between new modules, see [here](https://docs.ansible.com/ansible/latest/collections/ansible/builtin/include_module.html)